(split from Whiteboard) Merged PR 27318: feat: Add telemetry for invalid/malformed edits to SharedTree
Telemetry will be logged (via the logger supplied to SharedTree) when edits are invalid or malformed. Specifically, telemetry will be logged the first time an invalid/malformed local edit is applied by the LogViewer *after being sequenced*. It will be logged only by the client that originally generated the edit, and only the first time it is applied. This avoids redundant telemetry being logged by multiple clients or multiple times by the same client.