Add N/N-1 version testing for compat tests (#4818)
We hit a few issues recently where N/N-1 compatibility was broken despite N/N-2 being fine. Currently the compat tests are only run with N/N-2 to work better with the release process. The current process has the following steps, which mean N/N-1 testing will fail in the period between bumping the version on main and releasing the packages on the branch:

    Bump the package versions in main (client is bumped to N+1, and compat tests test against N and N-1; N at this point does not exist)
    Create release branch from the commit before bumping the package version
    Create releases for the packages in the release branch

Along with this change and the changes in bump-version to handle it, the steps can be adjusted as follows:

    NEW Publish a pre-release package for client from main
    Bump package versions in main. The new bump-version will change the new N-1 versions in end-to-end tests to use the prerelease version.
    Create release branch from the commit before bumping the package version
    Create releases for the packages in the release branch
    NEW Remove the pre-release dependencies in end-to-end tests and use the just-released version instead.

For adding another version to the tests, I refactored everything that created things from the old version into its own file (oldVersion.ts and oldVersion2.ts). Things that create things from the current version remain where they are. The oldVersion files are aware of the current version's classes so that they can create combinations of old/new runtimes/loaders, but they aren't aware of other old versions.

This was because trying to do everything dynamically lead to issues with the same class not being type-compatible across versions, leading to lots of casting that removes type safety that is very useful in this scenario. It's possible some sort of duck typing/generics system would get around that while enforcing type safety, but that involves adding extra interfaces in addition to the refactoring/lots of gobbledygook (though I think the refactoring here gets a lot of the way to doing this if needed). Adding an extra version (like N-3) involves copying one of the oldVersion sources, updating the package references, adding the types in oldVersionTypes, and pointing any tests that care to that new version file.