(split from Whiteboard) Merged PR 27395: build: update and audit deps
This does the subset of version updates that:
- Are possible due to peer deps (had to keep older copy-webpack-plugin)
- Are listed by 'npm outdated' in the root, or handled by running 'update-package-dependencies' (This means that possibly breaking updates requiring changing the actual version string in the package files for deps not listed in the root package.json were missed)

Additionally an 'npm audit fix' was applied, which claimed to fix 20 of 20 vulnerabilities.

This is far from a complete updating of deps, but it should at least pull in all bug fix and security releases for packages which we use '^' with. Breaking version updates (and updates of packages not using '^') only referenced in sub-packages should be all thats really missing.