Reorganize socket error handling not to lose handler (name of listener that generated an event) (#4907)
The core of the issue here is that OdspDocumentDeltaConnection.create does error translation and it loses information in the process as it throws away original message. Reorganizing code such that we do not need to do translation other than adding / changing properties (canRetry), but error is correctly (driver-specific!) generated in the first place and it's passed that way through other ways of communication (including "disconnected" handler).