(split from Whiteboard) Merged PR 27693: fix: BREAKING CHANGE: type of TreeNodeHandle.node
`TreeNodeHandle.node` is not a `TreeNodeHandle` but is in fact a `ChangeNode`. This PR corrects the type and also restores `ChangeNode` as the type returned by `Snapshot` helper methods, since it is impossible or needlessly complex to write helpers that generate a `TreeNode<TChild>` for an arbitrary `TChild` type.

Related work items: #51030