Pass docId to Container.load (#5066)
Currently there is no hint that Container.load() expects the id arg to be tenantId/docId, even though it throws the tenant away. This change simplifies that arg to just be the docId, which has the added benefit of better consolidating the steps to parse the resolved URL within the Loader's resolveCore().