Pass site url as callback parameter to enable fetching Push token against external tenants (#4918)
In order for user to load Fluid content from a file which is stored in external O365 tenant it is necessary to request OAuth token with external tenant explicitly specified as part of AAD authority that should issue a token.

StorageTokenFetcher callback signature includes siteUrl parameter which can be used to determine AAD tenant. We need similar capability for PushTokenFetcher callback. This is required because Push token must be able to establish collab session within the context of external tenant.