Improve port usage for tinylicious and unit test (#5145)
TCP hangs on to a port after it close for a period of time to avoid message being delivered after.
If we make a lot of HTTP request, that create a lot of TCP connection. The unit test is an example.

Also within tinylicious, even though that it is monolithic, it is still a bundling of multiple microservices. So call to historian is still done thru HTTP request even though they are the same process.

To avoid port exhaustion, set the HTTP connection to keep alive so that it can be reused, instead of closing and creating a new one.

The change set those flag globally for the unit test and tinylicious. We might want to consider making this more targeted and configurable in the future.

Related to issue #5098

Also fixed an instability in the leader test.

Note: I am able to run the mocha e2e test with --parallel --jobs 2 now. Not enabling it because there are still instability.