add data Processing error coercions (#5144)
Definitions (for the purpose of this discussion)
Data corruption: a scenario in which participants of a session are unable to partially or fully leverage session data that is believed (i.e. code/protocol expectations rather than user/experience expectation) to be persisted
Data loss: a scenario in which a user might expect data to be persisted but the protocol understands that persistence has not yet occurred (and may never occur)
Objective
We want to rapidly expose signals that can be used in attempts to detect data corruption.

Plan
The current indicator we are relying on is for "thrown" error values to have an errorType of "dataProcessingError" -- this is built into the DataProcessingError class.

While walking through the code with @anthony-murphy , he identified 3 key places where we might be able to expose a "plurality" of Processing issues -- and we throw DataProcessingError instances to do so.