Minimize involvement of Container in request headers (#5175)
This change proceeds with the effort to increase separation and layering between the Loader and Container (aligning with the pay-for-play idea of using Container without a Loader), by pulling LoaderHeader use out of the Container. The Container then expects more-explicit options to be set upon it, rather than inferring from the headers of the IRequest it is given. To further enforce this, Container no longer takes any IRequest arguments in its static load or constructor.