Fix flaky test deRehydrateContainerTests.spec.ts (#5259)
deRehydrateContainerTests.spec.ts has been flaky for a while. It is hard to repro, so not sure this will fix it.
But ensureSynchronized() is better then OpProcessingController in that it handle pending change better.
So hopefully this will fix the flakiness.

I ran the test repeatedly 100 times locally with no repro. (Where as I was able to repro it 9 out 100 times before)