Restrict ILoader interface (#5349)
From the discussion on the role of the ILoader/Loader on Tuesday.

Remove APIs on the ILoader interface that don't need to be there (around creating detached containers). Callers are expected to have a concrete Loader in cases they would want to call these because they require knowledge of the drivers. Move these APIs to a new IHostLoader interface, and have Loader implement that instead.

Most local changes are related to types in test code, and all instances where Bohemia calls the removed APIs are done on the Loader class rather than ILoader interface