(split from Whiteboard) Merged PR 28564: feat: Cache snapshots for local edits
This PR allows SharedTree to cache snapshots for local edits. This is accomplished by maintaining a queue of snapshots that is pushed/shifted as local edits are added/sequenced. The queue cache is invalidated (cleared) upon reception of a sequenced edit generated by a remote client. This change improves constant factors in edit application + sequencing substantially, and removes bad asymptotics in the handling of local edits which was noticeable in high-frequency (online) editing and dominant in offline scenarios.

This PR also adds a Queue data structure that allows for O(1), zero-allocation pushes/shifts/clears.

Related work items: #52562