Op fetch batch size: move to 2K to 5K as default, keeping r11s at 2K (#5361)
Closing #5211 - please see that issue for a bunch of details on timing, sizing, etc.
Long term, we will move to streaming API, leaving up to the driver to decide on strategy (fetch in chunks, or stream all ops in one call, etc.).

https://github.com/microsoft/FluidFramework/issues/5393 will be next step and tracks parallel requests.
After this is done, I'll start moving code out of DeltaManager into driver-util package for drivers to be able to use it, but switch driver API to be streaming.

Note, that with all the telemetry just added (see #5273, #5360), we will learn on the best strategy to catch up the fastest (given payloads we have). It's possible that I'll need to go back to 2K ops as it's faster to do 3x overlapping of 2K ops, vs. 1x 5K.