Add option to include Loader in scope (#5480)
Fixes #5329

Add a LoaderOption to add the created Loader into the scope object sent into the created Containers. This allows data objects to access a Loader without using the loader property on the context, and allows us to remove the loader property later (thus making loader access off by default).

This exists as a loader option at creation rather than having the host add the loader to the scope because 1) it standardizes the access pattern, and 2) the scope is passed into the Loader that wants to be added, but the scope is readonly on the Loader. Using a creation option seems easier for the consumer than creating a scope, passing it into the Loader, then editing the scope in the Loader after it has been created and forcing the data object to be aware of the property names the host is using.