Add disconnect/reconnect support to TaskManager (#5626)
The TaskManager is attempting to negotiate exclusive ownership of a task amongst connected clients. As a result, a disconnect should be interpreted as having lost all locks and left all queues. Even if outstanding "volunteer" ops are eventually ack'd, they will get reconciled by all clients once the "clientLeave" comes through. This change implements that behavior and adds tests for the same.

This change also updates the clicker demo with retry logic in the case that the task is lost or failed to acquire -- otherwise, the multiple connect/disconnects during normal load (connect, nack/reconnect) were causing the logic to fail since it was only trying a single time for the first connection. Now it will try to re-enter the queue if the lock is lost or after reconnect.