TaskManager only add valid clients (#5631)
Due to the way we delay load datastores and dds it's possible for them to be out of sync with the client until they are fully loaded. Specifically, they can see ops older than that latest quorum. The quorum will not contain the clients that these ops are from, as they have already been removed. This leads to zombie clients owning tasks in task manager

Trying to understand the situation better -- is this saying that we will miss the clientLeave for the zombie client in the scenario that it is processed before the TaskManager is loaded? And that the scrubClientsNotInQuorum call in loadCore comes at the wrong time to correct somehow?

exactly. we will load from the snapshot, but may have outstanding ops that are cached for the datastore/dss. these will be processed after we are loaded from snapshot, but are old in comparison to the quorum, so the clients could be gone by the time we get their op