Fix replay tool (#5610)
The non-decoded version should be as unaltered as possible so that we can just return what is read on disk. The decoded directory is only for human eyes only. Tools shouldn't read from it. 

This change fluid-fetch tool to write out the non-decoded blob as is (without encoding it in base64 first), and the replay tool revert back to reading those non-decoded blobs.

Currently the file document service read the blob from the "decoded" directory, and looking for file without the .json extension, which the fluid fetch tool always write to. So the replay-tool --from option is broken and this change fixes that.