(split from Whiteboard) Merged PR 28670: feat: implement opt-in blobed payload support in shared-tree and fluid-document-store
Adds support for blobbing payloads which opt in (by providing an asyncPayloadHandler). This will also async load the payloads, avoiding blocking on them when downloading.

Opting a definition into blobing and putting it into a document will mean the document will not work with versions of fluid-document-store preceding this change, so a version of the app supporting blobs should be deployed then a separate deployment should be done to opt-into blobbing so that a rollback can not result in documents which can not be loaded.

This feature supports synchronously loading blobs if they do not have a asyncPayloadHandler (they will be prefetched), once  on a version of fluid-document-store with this feature, there will be no document compatibility issues when changing which payloads get blobbed.

Related work items: #48025