Add Pairwise Options Exploration to Stress Test (#5697)
We're starting to leverage options to control more features and this increases the number of possible combinations of code interactions. This change introduces two major things:

Strongly type OptionsMatrix for options objects that will break the build if new options are add, but not included here. this will ensure we introduce new coverage for new features
Pairwise generation of options objects that are used for runners such that we have good coverage over a broad set interactions