ODSP Driver: Reduce Redundancy and Improve Consistency of Contracts (#5704)
This change was spurred by a missed property, fileVersion on OdspDocumentInfo which wasn't plumbed through. This contract is redundant and just add complexity to making changes. On fixing I identified a number of contract issues, so decided to do a bunch of cleanup that should make the api easier to use with each other, and reduce to the need to manually spread objects into parameters repeatedly. No fluid contracts are affected by this change, only helper functions may change