Remove Unused Legacy Container Hot Swap (#5725)
Removes all code related to hotswap, which has never been used, and doesn't work. To support legacy containers created before create detached, i added a wait in the protocol handler to wait until the code details has been set, this allows remove of the null runtime. The snapshot test cover this case. Lastly renamed the simplified reloadContextCore to processCodeProposal and added an error if the container is closed due to the current context not satisfying the incoming context

fixes #5471
fixes #3738
related #4685