FluidStatic Container API Re-design and DDS support (#5782)
This is a pretty major re-write to the FluidStatic Container API surface. This change has come from two major considerations. First is the addition of DDSes. Second is considering ways we can support client versioning in future iterations. The intention is that the API and patterns here will remain moving forward.

I recognize that this is a large change but didn't see value in staging it.

Some notable changes:

The README.md now has a basic description of usage (sans Audience)
New CHANGELOG.md captures breaking version changes
Breaking changes can be read in details in the CHANGELOG file.

Implementation changes that are not captured there:

FluidContainer is now an interface and we return the RootDataObject instead of wrapping the root in a proxy class.
Types are now being stored in ./src/types.ts.