(split from Whiteboard) Merged PR 29605: perf: BREAKING CHANGE: Improve performance of inserts/deletes/replaces in Forest
Currently, insertion/removal/replacement of a node in Forest could only be accomplished via a call to 'updateTraitContents' and causes all siblings (in all traits) under the parent node to be unparented and reparented. This was done in order to update cached indices and update parentage for any nodes effected (because 'updateTraitContents' can not efficiently diff the new trait contents with the old). This results in a large reduction in the effectiveness of the copy-on-write btree approach to Snapshots, as the churn in siblings caused much of the tree to be duplicated.

In this PR, this is alleviated by:
1. Removing APIs from Forest that suffer from this issue, namely `updateTraitContents`, `mergeWith`, and `replaceNode`.
2. Replace the above APIs with the more intentional `insertIntoTrait`, `detach`, and `replaceNodeDat`. These are all clear in the nodes they affect that require reparenting (the adds, the removes, and none, respectively).
3. Getting the index of a node is now a O(trait siblings) operation on Snapshot, but is cached. This PR is still an overall asymptotic improvement and is enormously beneficial to the copy-on-write pattern.
4. As a bonus, new unit tests of Forest, Snapshot, and Transaction were added. Usages of Forest in these unit tests now have expensive consistency asserts turned on.
5. Duplicate node IDs in the same Build op are now handled as Malformed instead of a crash.

Related work items: #54864