(split from Whiteboard) Merged PR 29483: refactor: BREAKING CHANGE: separate shared-tree code into edit-specific and edit-agnostic classes and types
Refactors the shared-tree package such that all types and code that depend on a specific set of changes are under a separate src/default-edits/ folder. None of the code in /src (excluding the code in /src/test/ and /src/default-edits/ and excluding src/index.ts) imports from src/default-edits/.

This makes SharedTree usable with a different set of edits, makes the package code more maintainable and easier to reason about.

This PR is the first of two PRs. The second will move generic types to a separate src/generic/ folder. The two-step process makes it easier to inspect the changes.

Related work items: #53110