Port https://github.com/microsoft/FluidFramework/pull/5773 to 0.38
Addresses #5477

It gives a lot of flexibility to developers to control loading process.
In particular, it allows Bohemia to apply cached ops before container is returned, and (optionally) allow ops from storage to continue to trickle in through initial boot / render cycle to speed up catch up process (this would need to be tested to ensure existing code behaves correctly in presence of changes).