(split from Whiteboard) Merged PR 29789: fix: Correctly use minimum sequence number in cache
An incorrect assumption had been made that, when loading a summary, numberOfSequencedEdits <= minimumSequenceNumber. This is false (ex: one client could summarize at revision 10, a new client could join and load that summary, but a client already in the session is still processing revision 5).  This fixes this bad assumption.

Related work items: #55367