(split from Whiteboard) Merged PR 29566: refactor: BREAKING CHANGE: move benchmarking code into its own package
This leaves the bench-marking code mostly as is, but moved to its own package.

Further PRs will rename and reorganize the code within the package to make more sense in its new context.

This is being done in preparation for offering this bench-marking functionality to the fluid-framework project for their use.

Related work items: #51962