(split from Whiteboard) Merged PR 29909: fix: BREAKING CHANGE: Expose factory for creating SharedTrees that do not preserve history in summary
Currently, the mechanism for preventing history from being saved in a summary is broken. This is because the factory that creates the summarizing client cannot be parameterized. The future fix for this is to use (not yet implemented) support from Fluid Framework to parameterize the DDS on first creation. This PR is a workaround that exposes a SharedTreeFactory flavor that paramaterizes the DDS instance after creation.

Related work items: #52686