Make Loader provide itself to Containers by default (#6032)
Tony opened a draft a bit ago (#5832) of what decoupling the Loader object from the Container might look like, and one takeaway was that in this scenario, if a Loader creates a Container, it makes sense to provide itself by default in the scope (because creating a Container without a Loader would also be an option). When integrating some loader changes to Bohemia, keeping track of all the places to explicitly provide the Loader on the scope got messy, so given there's an eventual goal of decoupling the Loader object from the Container once its constituent pieces are also decoupled, I think it makes sense to go ahead and adopt this part of Tony's change (along with some other bits of cleanup as well). Tony's change also removed the provideScopeLoader option; it's retained here but changed to true by default to provide hosts the option to override this behavior.