Set up automation for tinylicious-client (#6197)
Set up initial scaffolding for tinylicious-client tests that run using actual tinylicious in the same way as existing realsvc tests in end-to-end-tests, along with adding a basic basic test. This opens the path to adding additional tests as more functionality gets added here.

Version compat testing will be addressed in the future when APIs have been more solidified. There is also additional work to actually write that compat infrastructure/adapt the existing infra rom end-to-end-tests
This does not (and should not) run as part of PR validation or as part of any existing CI pipelines.
It will be added into the CI at some point in the future.
Because the tests run in mocha, I had to adjust the tsconfigs for tinylicious-client and fluid-static to not compile to esnext modules for the dist output folder (it was probably a copy-paste mistake that this was the case in the first place)