reverting #6041
Reverting PR #6041 "Updated disabling tenant to tenant deletion."

Reason: when we delete tenant, we also need to delete the related tenant data like operations (in mongoDb) and summaries (in git). So, we will take up this work along with a larger work item to delete the tenant data (both, mongodb and git) as part of delete tenant.