Test that rehydrates serialized container contents checked into repo (#6218)
The proposal is to check in a test that verifies that a specific container can be successfully rehydrated. The goal is to make explicit backwards compatibility breaks, such as (I believe) this one.

Note that at any specific point in the future, it may or may not make sense to break backwards compatibility, depending on guarantees to and requirements of downstream partners, which are likely to evolve.