Fix reset for test-service-load runner (#6366)
Since the iteration is currently happening at the top of the loop, the reset condition is never met (it originally would reset on the first iteration of the loop only). This change restores the reset behavior, and also tweaks the loop a bit to make the reset and loop exit conditions more obvious to a new reader of the code.