(split from Whiteboard) Merged PR 30719: feat: Anchor resolution
## Overview

This PR introduces a new set of edits in `src/anchored-edits`. These are the same as the edits from `src/default-edits` except when it comes to merge resolution: anchored edits are expressed in terms of anchors (instead of tree locations) which provide greater flexibility in specifying merge semantics.

This PR adds exactly one set of new semantics (see `PlaceAnchorSemanticsChoice.RelativeToNode`) instead of the one choice of semantics previously offered through `src/default-edits` (which is still made available in `src/anchored-edits` through `PlaceAnchorSemanticsChoice.BoundToNode`).

## Concepts

### Anchor

An anchor is a tree location augmented with information that describes how the anchor should be resolved in the advent of concurrent changes.

### Reconciliation Path
When an edit is applied, there is a chance that it may be applied to a snapshot that is different from the snapshot for which it was originally issued. This happens because other (concurrent) edits might have been sequenced first. The reconciliation path is the theoretical set of changes (and associated snapshots) that would need to be applied to the snapshot for which the edit was issued to arrive at the snapshot state at which the edit is actually applied.

In this PR, we introduce this concept and compute only the "forward" portion of it (i.e., the edits on the main branch starting at the last state common two both edit branches). The "backward" portion of it is needed for the sequential merge approach which is not supported yet.

This reconciliation path information is fed to the edit application logic so that it may take into account the changes that have occurred since the last state in common to both branches.

### Resolved Change
A resolved change is a change whose anchors have been resolved to valid tree locations.

## Known Issues
As of this PR, SharedTreeWithAnchors does not guarantee convergence among nodes.
This is due to the fact that the current summary information is insufficient to ensure all nodes have access to all the relevant data for the application of a given edit. For example, if edit was sequenced after a given summary but references an edit before that summary then nodes that do not have that earlier edit's information in memory will not have the full context for the application of the change. This currently leads to a "best effort" application which takes into account whatever data is available.
This is tracked by #57176.

## Changes
* `GenericSharedTree.processCore` now reads server-assigned sequence information from incoming edits.

* `GenericTransaction.applyChange` and `GenericTransaction.applyChanges` now take the following additional parameter: `path?: ReconciliationPath<TChange>`

* `GenericTransaction.applyChange` now calls the virtual method `GenericTransaction.tryResolveChange` which allows child classes to implement anchor resolution.

* `GenericTransaction` now keeps track of (and returns on `close()` via `ValidEditingResult...