MergeTree: Record Conflict Farm Results and Enable Replay Tests (#6381)
This is related to #1214 . I have a potential fix, but am concerned about behavior change, to mitigate this concerns i've modified the conflict farm stress test such that the results can be recorded.
The recorded results include the initial text, the operations, and the result text for each group of operations. With these i've also added a replay test that can replay these files and ensure that starting from an initial state and applying operations results in the matching end state.
This helps cover a gap where we primarily validate that all clients converge to the same state and didn't validate that that converged state itself was consistent. This will also give us confidence that that clients running before and after version of the #1214 fix will reach the same end state in our currently covered scenarios.

Co-authored-by: Tony Murphy <anthonm@microsoft.com>