Merge Tree: Differentiate removed segments, and not yet visible segments (#6386)
This change fixes a long standing but rare issue in the merge tree. The issue was caused by an inability in the insertion logic to disambiguate between removed segments, and no yet visible segments. Specifically in both cased node length would return a value of 0, and all 0 length segments were eligible for conflict resolution if targeting the same position. Also not all clients necessarily have the same removed segments as removal of removed segments is asynchronous process (Zamboni). These facts could lead to one client breaking a tie on some removed segment, whereas other clients would not, and this would lead to incorrect ordering of conflicting segments.

The fix is to report removed segments that will never be visible as undefined length, and to ignore those segments on insert, and only considers segments that will be visible in the future (0 length segments) for conflict resolution.

This change should go in after #6381 as it includes validation for this change.

Fixes #1214