Improve mechanism for wrapping errors e.g. createContainerError (#6477)
We have code that takes an arbitrary `error` and converts it to a more canonical error object with a meaningful type and also capabilities like safe logging of props.  This was implemented a few different places with slightly different semantics - and some bugs (e.g. the pitfall of #6468). This PR unifies this to a single `wrapError` function that takes a callback allowing for wrapping in different types of errors, but sharing the core logic / semantics.