Remove `existing` property - Part 4, IRuntimeFactory + RuntimeFactoryHelper  (#6604)
See #3429

optional argument for IRuntimeFactory.instantiateRuntime to replace the existing state
RuntimeFactoryHelper is to be used by current clients of IRuntimeFactory (see RFC - remove use of the existing property across (some) layers #6486 for examples) in a future PR, after the version bump in container definitions. This class now holds the wrapper logic about how the runtime is actually instantiated.

* Support existing in IRuntimeFactory, add abstract helper class for IRuntimeFactory

* Add test infra

* Add small test

* Simplify mocks

* Use partials to mock IContainerContext, add more tests

* Rename tests, make precedence test more explicit

* Remove comment

* Add some param comments

* Update api.md

* remove marker property

* Remove redundant check

* Some PR feedback, make abstract class generic

* Type refactoring

* Fix test

* Fix test typing

* Fix cast