Cherry-pick #6604 from main. IRuntimeFactory signature change (#6638)
* Remove `existing` property - Part 4, IRuntimeFactory + RuntimeFactoryHelper  (#6604)

See #3429

optional argument for IRuntimeFactory.instantiateRuntime to replace the existing state
RuntimeFactoryHelper is to be used by current clients of IRuntimeFactory (see RFC - remove use of the existing property across (some) layers #6486 for examples) in a future PR, after the version bump in container definitions. This class now holds the wrapper logic about how the runtime is actually instantiated.