Deprecate IErrorBase.sequenceNumber (#6639)
This field was used for logging and this was probably not the right abstraction for it to live in.

We can stop using it because the only places it was set have been updated to log not just sequenceNumber
but a large number of useful properties off the offending message, via `CreateProcessingError`.

Note that `ContainerClose` event will now have a field `lastSequenceNumber` which will be from `DeltaManager`, and in cases where an op processing error occurred, `messageSequenceNumber` will have the same value `sequenceNumber` used to.  See [here](https://github.com/microsoft/FluidFramework/blob/c32fff7c6214a8562bb6cf416eed94a6eb502a4f/packages/loader/container-utils/src/error.ts#L151) for the full list of op properties logged.