Add a unique ID to Interval properties, allow multiple intervals with the same (start, end) (#6407)
* Add methods to get and remove an interval from a collection by ID

Add and delete ops will be supplied with ID's. delete(start, end) will be deprecated; for now it will delete the interval with a legacy ID based on start/end.

* Fix behavior of delete(start, end) and a cross-client issue with transmission of ID's

* Allow multiple intervals with the same (start, end) in a collection

Use the interval ID as a third key for sorting. Fix an issue in serialization of add ops that was causing the original ID not to be sent over the wire.

* Add dehydrate+rehydrate test

* Restore global uniqueness of legacy ID's