Event improvements and refactoring for ServiceAudience (#6673)
Some QoL changes to ServiceAudience events to make them easier to work with.

Split the membersChanged event back into memberAdded and memberRemoved and provide both the clientId and IMember object for the affected member in the event. For consumers who track data about audience members in the user -> connections -> data format (which is the same way ServiceAudience formats its member information), this makes it easier to make incremental changes to that data instead of needing to compare all ServiceAudience members with all user-known members to determine the affected member as was needed before.
Change event behavior to only emit for member changes that are reflected in the member list provided by the ServiceAudience. This means a summarizer client leaving IAudience doesn't trigger ServiceAudience events because there is no externally visible change to ServiceAudience members.
Do some generics stuff on ServiceAudience and make it abstract so child classes duplicate less code
Update FocusTracker example to reflect changes to ServiceAudience