Add a signal manager class for fluid-static to help with signal usages (#6659)
Add a signal manager class for fluid-static that can wrap an existing signaler, and integrate into the focus tracker example to demonstrate how that would look. The signal manager exposes the signal API outside of the class. It presents signals more closely to eventing and makes it easier to build into the listener/broadcaster pattern that signal users usually adopt. Managers can also keep an internal id that is applied to signal names, so users don't need to worry about name conflicts at the container level.

For focus-tracker, we see that while the helper does not reduce the amount of code, the way signals are used is more centralized and easier to understand. The built-in broadcast pattern means signal users also no not need to worry as much about connection state/audience edge cases that might sometimes result in clients not sending or being able to receive a signal at the right time.