(split from Whiteboard) Merged PR 32352: test: Summarizer Read Forward Compatibility Tests
This PR adds a generalized testing format for summarizer read forward compatibility tests. The specialized test for summarizer version `0.0.2` on summaries with handles is changed to be included as part of these generalized tests. As future summarizer versions are added, a `ForwardCompatibilityTestEntry` should be added to `forwardCompatibilityTests`.

Related work items: #60277