Fix crash in server tests on windows (#6691)
Loading the node-rdkafka module without create an instance of RdKafka would cause the node process to AV when exiting (observed in windows). This causes our npm run test command to fail even though all the test passed.

The solution is to only load the module if we are using RdKafka.