Add signal manager data object for fluid static (#6797)
This already came up before in an earlier PR, but was converted to a standalone class. After some discussion, we decided providing it as a DO was a preferable starting point over adding it as a first class API on the FluidContainer class (or adding it as part of container services because those are expected to be service specific). The reasoning being we must do one of the 3 (or some variant thereof) because we are not providing a way to use the standalone class as part of v1, and because it's much easier to promote the SignalManger to a first class API later if needed than the other way around.

We do not expect developers to create lots of these (and agitate issues related to its being a DO), but we will keep it in mind as part of observing usage to decide if it needs to change.